+++ src/debops/config.py
@@ -1,7 +1,7 @@
 # -*- coding: utf-8 -*-
 
-# Copyright (C) 2020 Maciej Delmanowski <drybjed@gmail.com>
-# Copyright (C) 2020 DebOps <https://debops.org/>
+# Copyright (C) 2020-2023 Maciej Delmanowski <drybjed@gmail.com>
+# Copyright (C) 2020-2023 DebOps <https://debops.org/>
 # SPDX-License-Identifier: GPL-3.0-or-later
 
 from .utils import unexpanduser
@@ -27,6 +27,7 @@ class Configuration(object):
     def __init__(self):
 
         self._env_files = []
+        self._env_vars = {}
 
         # Include variables from the system-wide configuration
         self.merge_env(os.path.join('/etc', 'default', 'debops'))
@@ -90,6 +91,7 @@ class Configuration(object):
 
     def set_env(self, key, value):
         os.environ[key] = str(value)
+        self._env_vars[key] = str(value)
 
     def get(self, items=None):
         if items:
@@ -110,10 +112,12 @@ class Configuration(object):
             if os.path.isfile(os.path.join(path, '.env')):
                 self._env_files.append(os.path.join(path, '.env'))
                 dotenv.load_dotenv(os.path.join(path, '.env'), override=True)
+                self._env_vars.update(dotenv.dotenv_values(os.path.join(path, '.env')))
         elif os.path.exists(path):
             if os.path.isfile(path):
                 self._env_files.append(path)
                 dotenv.load_dotenv(path, override=True)
+                self._env_vars.update(dotenv.dotenv_values(os.path.join(path)))
 
     def load(self, path):
         if os.path.exists(path):
@@ -121,7 +125,7 @@ class Configuration(object):
                 return self._load_config_file(path)
             elif os.path.isdir(path):
                 data = {}
-                for config_file in os.listdir(path):
+                for config_file in sorted(os.listdir(path)):
                     if os.path.isfile(os.path.join(path, config_file)):
                         self._merge_dict(data,
                                          self._load_config_file(
@@ -200,25 +204,73 @@ class Configuration(object):
                              .update({key: value}))
             return self._converted_data
 
-    def show_env(self):
-        for key, value in os.environ.items():
-            print('{}={}'.format(key, value))
+    def config_list(self):
+        relative_root = os.path.relpath(os.path.abspath('/'))
+        if self._env_files:
+            print('# Environment files:')
+            for filename in self._env_files:
+                relative_file = os.path.relpath(unexpanduser(filename))
+                print(relative_file.replace(relative_root, '', 1))
+            print()
+        if self._config_files:
+            print('# Configuration files:')
+            for filename in self._config_files:
+                relative_file = os.path.relpath(unexpanduser(filename))
+                print(relative_file.replace(relative_root, '', 1))
 
-    def show(self, format='toml'):
-        if format == 'toml':
-            relative_root = os.path.relpath(os.path.abspath('/'))
-            if self._env_files:
-                print('# Environment files:')
-                for filename in self._env_files:
-                    relative_file = os.path.relpath(unexpanduser(filename))
-                    print('#    ', relative_file.replace(relative_root, '', 1))
-                print()
-            if self._config_files:
-                print('# Configuration files:')
-                for filename in self._config_files:
-                    relative_file = os.path.relpath(unexpanduser(filename))
-                    print('#   ', relative_file.replace(relative_root, '', 1))
-                print()
-            print(toml.dumps(self._config).strip())
-        elif format == 'json':
-            print(json.dumps(self._config, sort_keys=True, indent=4))
+    def config_env(self, scope='local'):
+        if scope == 'local':
+            for key, value in self._env_vars.items():
+                print('{}={}'.format(key, value))
+        elif scope == 'full':
+            for key, value in os.environ.items():
+                print('{}={}'.format(key, value))
+
+    def config_get(self, key, format='unix'):
+        key_path = ['.']
+        if key != '.':
+            key_path = list(filter(None, key.split('.')))
+        key_found = True
+
+        _config = self._config
+        if key_path[0] != '.':
+            for element in key_path:
+                try:
+                    _config = _config[element]
+                except KeyError:
+                    key_found = False
+
+        if key_found:
+            if isinstance(_config, dict):
+                if format in ['unix', 'yaml']:
+                    print(yaml.dump(_config).strip())
+                elif format == 'json':
+                    print(json.dumps(_config, sort_keys=True, indent=2))
+                elif format == 'toml':
+                    print(toml.dumps(_config).strip())
+            elif isinstance(_config, bool):
+                if format == 'unix':
+                    print(json.dumps(_config))
+                elif format == 'yaml':
+                    print(yaml.dump(_config).strip())
+                elif format in ['json', 'toml']:
+                    print(json.dumps(_config))
+            elif isinstance(_config, list):
+                if format == 'unix':
+                    for entry in _config:
+                        print(entry)
+                elif format == 'yaml':
+                    print(yaml.dump(_config).strip())
+                elif format in ['json', 'toml']:
+                    print(json.dumps(_config, sort_keys=True))
+            else:
+                if format == 'unix':
+                    print(_config)
+                elif format == 'yaml':
+                    print(yaml.dump(_config).strip())
+                elif format in ['json', 'toml']:
+                    print(json.dumps(_config))
+        else:
+            if format == 'json':
+                print('{}')
+            sys.exit(1)